added the get, delete and update by id in the try block

+++ warsan/api/views.py
@@ -8,17 +8,16 @@ from .serializers import VaccineSerializer
 
 class VaccineListView(APIView):
     def get(self, request):
-        vaccine=Vaccine.objects.all()
-        serializer=VaccineSerializer(vaccine, many = True)
+        vaccines = Vaccine.objects.all()
+        serializer = VaccineSerializer(vaccines, many=True)
         return Response(serializer.data)
-    
-    def post(self,request):
+
+    def post(self, request):
         serializer = VaccineSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
 
 class VaccineDetailView(APIView):
     def get_object(self, pk):
@@ -27,35 +26,32 @@ class VaccineDetailView(APIView):
         except Vaccine.DoesNotExist:
             raise Http404
 
-    def get(self, request, pk):
+    def handle_vaccine_operations(self, request, pk, operation):
         try:
             vaccine = self.get_object(pk)
-            serializer = VaccineSerializer(vaccine)
-            return Response(serializer.data)
-        except Http404:
-             return Response("Vaccine not found", status=status.HTTP_404_NOT_FOUND)
-
-
-    def put(self, request, pk):
-        try:
-            vaccine = self.get_object(pk)
-            serializer = VaccineSerializer(vaccine, data=request.data)
-            if serializer.is_valid():
-               serializer.save()
-               return Response(serializer.data)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        except Http404:
-           return Response("Vaccine not found", status=status.HTTP_404_NOT_FOUND)
-
-
-    def delete(self, request, pk):
-        try:
-            vaccine = self.get_object(pk)
-            vaccine.delete()
-            return Response("Vaccine Deleted", status=status.HTTP_204_NO_CONTENT)
+            if operation == 'get':
+                serializer = VaccineSerializer(vaccine)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            elif operation == 'put':
+                serializer = VaccineSerializer(vaccine, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return Response(serializer.data, status=status.HTTP_200_OK)
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            elif operation == 'delete':
+                vaccine.delete()
+                return Response("Vaccine Deleted", status=status.HTTP_204_NO_CONTENT)
         except Http404:
             return Response("Vaccine not found", status=status.HTTP_404_NOT_FOUND)
 
+    def get(self, request, pk):
+        return self.handle_vaccine_operations(request, pk, 'get')
+
+    def put(self, request, pk):
+        return self.handle_vaccine_operations(request, pk, 'put')
+
+    def delete(self, request, pk):
+        return self.handle_vaccine_operations(request, pk, 'delete')
